+++ test/unit/builtinEvents.js
@@ -1,7 +1,7 @@
 import test from 'ava'
 import sinon from 'sinon'
 import Vue from 'vue/dist/vue.common'
-import VueTouch from '../src/index.js'
+import VueTouch from './helpers/vue-touch'
 import Hammer from 'hammerjs'
 
 Vue.use(VueTouch, {hammer: Hammer})
