+++ test/unit/customEvents.js
@@ -1,7 +1,7 @@
 import test from 'ava'
 
 import Vue from 'vue/dist/vue.common'
-import VueTouch from '../src/index.js'
+import VueTouch from './helpers/vue-touch'
 import Hammer from 'hammerjs'
 
 VueTouch.registerCustomEvent('doubletap', {
